+++ HISTORY.rst
@@ -79,6 +79,7 @@ End-User Summary
 - Renamed all occurrences of whitelist to allowlist and of blacklist to blocklist (sticking to what google introduced in their products).
 - Fixed bug where cases were not deletable when using Chrome browser.
 - Harmonized computation for relatedness in project-wide QC and in case QC (thus showing the same results if project only contains one family).
+- Fixed failing case API re-import when user is not owner of previous import.
 
 Full Change List
 ================
@@ -169,6 +170,8 @@ Full Change List
 - Renamed all occurrences of whitelist to allowlist and of blacklist to blocklist (sticking to what google introduced in their products).
 - Fixed bug where cases were not deletable when using Chrome browser.
 - Harmonized computation for relatedness in project-wide QC and in case QC (thus showing the same results if project only contains one family).
+- Fixed failing case API re-import when user is not owner of previous import.
+  Now also all users with access to the project (except guests) can list the cases.
 
 -------
 v0.21.0